[radio] Initial commit (core package)
This commit adds the core package that will power the 'radio' service
which is responsible for (email) communication. This service exposes a
JSON api (over http) that callers on the box can use to send notifications
to the world. The service has a notion of message severities and callers
need only specify a severity so that radio can determine who needs to be
notified.

A subsequent commit will add the server code that uses this package.