[crashcorder] Add crashcorder core package
This commit adds the crashcorder package that provides the functionality
to watch a particular directory for core files and report them to the
radio service so that users are notified of process crashes. It also
sets up the kernel configuration so that cores are written to the
directory being watched.

A subsequent commit will add the service that uses this package.