[commander] Fix incorrect group IDs
The rootfs on which the rocketship system is based ships with certain
dirs having their permissions set to IDs from the upstream /etc/group
file. In our case, we try to mimic that as far as possible so that we
do not have to twiddle dir/file permissions later in the build or at
run time (though eventually we may want to control all of that as well).

For now, update our group->gid map to be in sync with 14.04 rootfs.